big changes coming
Removed all warnings, I didn't like them anymore and the library size is now much smaller
Color changes, created a ColorUtils
Worked on webpack output, now it's much smaller and used tsc instead of babel. Probably switched to ES6, but ES5 wasn't a requirement for me
Add pathseg dependency, now it's required to be inclued externally and this change will be documented, before it was included in the js bundle.
Removed support.js and its type file, it's was used only for pathseg require.