feat(index): upgrade to React 18
Installing the React app using the default TS-Particle templates I had faced following warnings in the browser console. This is because the templates the code inside the src/index.js is still of REACT@17+. But the installed react package is @18+.

Therefore I have updated the src/index.js code to the lastest as per REACT@18+ docs.